@@ -24,33 +24,33 @@
         "package:publish": "standard-version && git push --follow-tags origin main && npm publish"
     },
     "devDependencies": {
-        "@playwright/test": "^1.46.0",
-        "@sveltejs/adapter-auto": "^3.2.4",
-        "@sveltejs/kit": "^2.5.22",
-        "@sveltejs/vite-plugin-svelte": "^3.1.1",
+        "@playwright/test": "^1.48.2",
+        "@sveltejs/adapter-auto": "^3.3.1",
+        "@sveltejs/kit": "^2.7.5",
+        "@sveltejs/vite-plugin-svelte": "^4.0.0",
         "@types/eslint": "9.6.0",
-        "@typescript-eslint/eslint-plugin": "^8.1.0",
-        "@typescript-eslint/parser": "^8.1.0",
+        "@typescript-eslint/eslint-plugin": "^8.13.0",
+        "@typescript-eslint/parser": "^8.13.0",
         "autoprefixer": "^10.4.20",
-        "dayjs": "^1.11.12",
-        "eslint": "^9.9.0",
+        "dayjs": "^1.11.13",
+        "eslint": "^9.14.0",
         "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-svelte": "^2.43.0",
-        "postcss": "^8.4.41",
+        "eslint-plugin-svelte": "^2.46.0",
+        "postcss": "^8.4.47",
         "postcss-load-config": "^6.0.1",
         "prettier": "^3.3.3",
-        "prettier-plugin-svelte": "^3.2.6",
-        "prettier-plugin-tailwindcss": "^0.6.6",
-        "prisma": "^5.18.0",
-        "svelte": "^4.2.18",
-        "svelte-check": "^3.8.5",
-        "svelte-lib-helpers": "^0.4.7",
-        "svelte-meta-tags": "^3.1.2",
-        "tailwindcss": "^3.4.9",
-        "tslib": "^2.6.3",
-        "typescript": "^5.5.4",
-        "vite": "^5.4.0",
-        "vitest": "^2.0.5"
+        "prettier-plugin-svelte": "^3.2.7",
+        "prettier-plugin-tailwindcss": "^0.6.8",
+        "prisma": "^5.21.1",
+        "svelte": "^5.0.0",
+        "svelte-check": "^4.0.0",
+        "svelte-lib-helpers": "^0.4.16",
+        "svelte-meta-tags": "^3.1.4",
+        "tailwindcss": "^3.4.14",
+        "tslib": "^2.8.1",
+        "typescript": "^5.6.3",
+        "vite": "^5.4.10",
+        "vitest": "^2.1.4"
     },
     "type": "module",
     "peerDependencies": {
@@ -62,18 +62,18 @@
         "@svelte-plugins/datepicker": "^1.0.9",
         "@types/node-steam-openid": "^1.0.3",
         "date-fns": "^3.6.0",
-        "express-session": "^1.18.0",
+        "express-session": "^1.18.1",
         "fastify": "^4.28.1",
         "flag-icons": "^7.2.3",
-        "flowbite": "^2.5.1",
-        "flowbite-svelte": "^0.46.15",
-        "flowbite-svelte-icons": "^1.6.1",
-        "mongodb": "^6.8.0",
-        "mongoose": "^8.5.2",
+        "flowbite": "^2.5.2",
+        "flowbite-svelte": "^0.46.23",
+        "flowbite-svelte-icons": "^1.6.2",
+        "mongodb": "^6.10.0",
+        "mongoose": "^8.8.0",
         "node-steam-openid": "^1.2.3",
         "passport": "^0.7.0",
         "passport-steam": "^1.0.18",
-        "tailwind-merge": "^2.5.2"
+        "tailwind-merge": "^2.5.4"
     },
     "keywords": [
         "svelte",
