@@ -1,5 +1,9 @@
 <script lang="ts">
     import type { PageData } from './$types';
     
-    export let data: PageData;
+    interface Props {
+        data: PageData;
+    }
+
+    let { data }: Props = $props();
 </script>
