@@ -8,20 +8,24 @@
 
     type TimeSlot = keyof typeof timeslots;
 
-    export let timeslot: TimeSlot = 'Last 7 days';
-    export let timeslots = {
+    interface Props {
+        timeslot?: TimeSlot;
+        timeslots?: any;
+    }
+
+    let { timeslot = $bindable('Last 7 days'), timeslots = {
         Yesterday: -1,
         Today: 0,
         'Last 7 days': 7,
         'Last 30 days': 30,
         'Last 90 days': 90
-    };
+    } }: Props = $props();
 
     let timeslots_keys: TimeSlot[] = Object.keys(timeslots) as TimeSlot[];
 
     let today = dayjs();
-    $: start = today.subtract(timeslots[timeslot], 'days').format('ll');
-    $: end = timeslot == 'Yesterday' ? start : today.format('ll');
+    let start = $derived(today.subtract(timeslots[timeslot], 'days').format('ll'));
+    let end = $derived(timeslot == 'Yesterday' ? start : today.format('ll'));
 </script>
 
 <div class="font-normal">
@@ -29,15 +33,17 @@
         class="mt-0.5 inline-flex gap-1 rounded-lg p-2 text-center text-sm font-medium text-gray-500 hover:text-gray-900 dark:text-gray-400 dark:hover:text-white"
         >{timeslot} <ChevronDownOutline size="lg" /></button>
     <Dropdown class="min-w-48">
-        <div slot="header" role="none">
-            <DropdownItem class="truncate text-gray-900 dark:text-white" href="#">
-                {#if start == end}
-                    {start}
-                {:else}
-                    {start} - {end}
-                {/if}
-            </DropdownItem>
-        </div>
+        {#snippet header()}
+                <div  role="none">
+                <DropdownItem class="truncate text-gray-900 dark:text-white" href="#">
+                    {#if start == end}
+                        {start}
+                    {:else}
+                        {start} - {end}
+                    {/if}
+                </DropdownItem>
+            </div>
+            {/snippet}
 
         {#each timeslots_keys as slot}
             <DropdownItem class="font-normal" href="#" on:click={() => (timeslot = slot)}>
@@ -45,8 +51,10 @@
             </DropdownItem>
         {/each}
 
-        <div slot="footer" role="none">
-            <DropdownItem class="font-normal" href="#">Custom...</DropdownItem>
-        </div>
+        {#snippet footer()}
+                <div  role="none">
+                <DropdownItem class="font-normal" href="#">Custom...</DropdownItem>
+            </div>
+            {/snippet}
     </Dropdown>
 </div>
