@@ -2,15 +2,19 @@
     import { ChevronRightOutline } from 'flowbite-svelte-icons';
     import { twMerge } from 'tailwind-merge';
 
-    export let title: string = '';
-    export let href: string = '#';
 
-    export let flat: boolean = false;
+    interface Props {
+        title?: string;
+        href?: string;
+        flat?: boolean;
+    }
 
-    $: aClass = twMerge(
+    let { title = '', href = '#', flat = false }: Props = $props();
+
+    let aClass = $derived(twMerge(
         'inline-flex items-center text-xs font-medium text-primary-700 sm:text-sm dark:text-primary-500',
         flat ? '' : 'rounded-lg p-2 hover:bg-gray-100 dark:hover:bg-gray-700 uppercase'
-    );
+    ));
 </script>
 
 <div class="flex-shrink-0">
